server: Use multiple forks to handle incoming request
A simple mechanism for creating N tasks for each incoming request and replying
with the one which finishes first. Assume we would have a problem to solve and
would like to run several different algorithms at the same time and respond as
soon as any one of them has a solution, this would do the trick. Mainly serves
as a test for how node.js clustering and cluster communication works.

We need to also use coffeescript compilation instead of direct running to make
forking work.

This might be related (not sure):

  https://github.com/jashkenas/coffee-script/issues/2786

Signed-off-by: Aki Saarinen <aki@akisaarinen.fi>